[MME/AMF] Fixed crash following Handover Request (#3014)
1. HandoverRequired
2. HandoverRequest
3. HandoverFailure
4. UEContextReleaseCommand
5. HandoverPreparationFailure

If UEContextReleaseComplete is not received,
the Source-UE will have the Target-UE.

6. HandoverRequired

There may be cases where the Source UE has a Target UE
from a previous HandoverRequired process. In this case,
it is recommended to force the deletion of the Target UE information
when receiving a new HandoverRequired.

7. HandoverRequest
8. HandoverFailure
9. UEContextReleaseCommand
10. UEContextReleaseComplete
11. HandoverPreparationFailure

... Crashed ...

+++ lib/ngap/build.h
@@ -29,7 +29,7 @@ extern "C" {
 #endif
 
 ogs_pkbuf_t *ogs_ngap_build_error_indication(
-    uint32_t *ran_ue_ngap_id,
+    uint64_t *ran_ue_ngap_id,
     uint64_t *amf_ue_ngap_id,
     NGAP_Cause_PR group, long cause);
 
@@ -39,7 +39,7 @@ ogs_pkbuf_t *ogs_ngap_build_ng_reset(
 
 void ogs_ngap_build_part_of_ng_interface(
     NGAP_UE_associatedLogicalNG_connectionList_t *partOfNG_Interface,
-    uint32_t *ran_ue_ngap_id, uint64_t *amf_ue_ngap_id);
+    uint64_t *ran_ue_ngap_id, uint64_t *amf_ue_ngap_id);
 
 ogs_pkbuf_t *ogs_ngap_build_ng_reset_ack(
     NGAP_UE_associatedLogicalNG_connectionList_t *partOfNG_Interface);
